+++ js/src/ui/index.js
@@ -46,6 +46,7 @@ export Page from './Page';
 export ParityBackground from './ParityBackground';
 export Portal from './Portal';
 export QrCode from './QrCode';
+export QrScan from './QrScan';
 export SectionList from './SectionList';
 export SelectionList from './SelectionList';
 export ShortenedHash from './ShortenedHash';
